requirements.txt

psycopg2-binary

pandas
scikit-learn
lightgbm
numpy

rank_bm25
pymorphy2

requests
beautifulsoup4
lxml

tqdm
joblib

jupyter
ipykernel

apache-airflow==2.8.0
apache-airflow-providers-postgres==5.8.0
apache-airflow-providers-elasticsearch==4.5.0
